refactor: Se refactorizo el router

+++ src/app/models/CategoryModel.php
@@ -31,10 +31,17 @@ class CategoryModel {
         return $query->fetchAll();
     }
 
-    /**Traigo registros mediante el ID para mostrar el form.edit precargado */
     public function getById(string $id): ?object {
         $query = $this->connection->prepare("SELECT * FROM category WHERE category_id = ?");
         $query->execute([$id]);
+
+        $result = $query->fetch();
+        return $result == false ? null : $result;
+    }
+
+    public function getByIdAndUserId(string $id, int $userId): ?object {
+        $query = $this->connection->prepare("SELECT * FROM category WHERE category_id = ? AND user_id = ?");
+        $query->execute([$id, $userId]);
 
         $result = $query->fetch();
         return $result == false ? null : $result;
@@ -46,12 +53,26 @@ class CategoryModel {
 
         $result = $query->fetch();
         return $result == false ? null : $result;
-    }   
+    }
+    
+    public function getByIdAndUserIdWithDescription(string $id, int $userId): ?object {
+        $query = $this->connection->prepare("SELECT description FROM category WHERE category_id = ? AND user_id = ?");
+        $query->execute([$id, $userId]);
 
-    /**Traigo solo el nombre mediante el ID de la categoria seleccionada */
+        $result = $query->fetch();
+        return $result == false ? null : $result;
+    }
+
     public function getByIdWithName(string $id): array {
         $query = $this->connection->prepare("SELECT name FROM category WHERE category_id = ?");
         $query->execute([$id]);
+
+        return $query->fetchAll();
+    }
+
+    public function getByIdAndUserIdWithName(string $id, int $userId): array {
+        $query = $this->connection->prepare("SELECT name FROM category WHERE category_id = ? AND user_id = ?");
+        $query->execute([$id, $userId]);
 
         return $query->fetchAll();
     }
@@ -66,16 +87,19 @@ class CategoryModel {
 
     public function add(string $name, string $nameId, string $description, string $type, int $userId): void {
         $query = $this->connection->prepare("INSERT INTO category (name, name_id, description, type, user_id) VALUES(?,?,?,?,?)");
+
         $query->execute([$name, $nameId, $description, $type, $userId]);
     }
 
-    public function delete(string $id): void {
-        $query = $this->connection->prepare("DELETE FROM category WHERE category_id = ?");
-        $query->execute([$id]);
+    public function deleteByIdAndUserId(string $id, int $userId): void {
+        $query = $this->connection->prepare("DELETE FROM category WHERE category_id = ? AND user_id = ?");
+
+        $query->execute([$id, $userId]);
     }
 
     public function update(string $id, string $name, string $nameId, string $description, string $type): void {
         $query = $this->connection->prepare("UPDATE `category` SET `name` = ?, `name_id` = ?, `description` = ?, `type` = ? WHERE `category`.`category_id` = ?");
+
         $query->execute([$name, $nameId, $description, $type, $id]);
     }
 }
